restructure of folder tree
